pyproject.toml
[tool.poetry]
name = "trainer"
version = "0.1.0"
description = ""
authors = ["Nurlan-Aliev <nualiev612@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["all"], version = "^0.116.1"}
uvicorn = "^0.35.0"
sqlalchemy = "^2.0.41"
pyjwt = "^2.10.1"
bcrypt = "^4.3.0"
asyncpg = "^0.30.0"
greenlet = "^3.2.3"
alembic = "^1.16.4"
redis = "^6.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
